# docs/03-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: DGLEA Masonic Passport API
  version: 1.0.0
  summary: Formal API specification for the DGLEA Masonic Passport platform.
  description: |
    Workflow-explicit, permission-aware, server-enforced API for the DGLEA Masonic Passport.

    Design constraints:
    - Backend is authoritative for permissions and lifecycle transitions.
    - Verification workflow transitions are explicit endpoints.
    - District/lodge/member scope checks are mandatory server-side.
servers:
  - url: /api/v1
    description: Version 1 API
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user and issue tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated session context
      responses:
        '200':
          description: Session context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
  /me:
    get:
      tags: [Me]
      summary: Get current user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
...
  /members/{memberId}/passport-records:
    parameters:
      - $ref: '#/components/parameters/MemberId'
    get:
      tags: [PassportRecords]
      summary: List passport records for member
      parameters:
        - name: sectionId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PassportRecordStatus'
        - name: recordType
          in: query
          schema:
            $ref: '#/components/schemas/ItemType'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Member passport records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPassportRecordList'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    post:
      tags: [PassportRecords]
      summary: Create draft passport record
      description: Creates records in DRAFT status only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePassportRecordRequest'
      responses:
        '201':
          description: Draft record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportRecord'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /members/{memberId}/passport-summary:
    parameters:
      - $ref: '#/components/parameters/MemberId'
    get:
      tags: [PassportRecords]
      summary: Brother passport summary
      responses:
        '200':
          description: Passport summary payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrotherPassportSummary'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /passport-records/{recordId}:
    parameters:
      - $ref: '#/components/parameters/RecordId'
    get:
      tags: [PassportRecords]
      summary: Get passport record
      responses:
        '200':
          description: Passport record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportRecord'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags: [PassportRecords]
      summary: Update draft passport record
      description: |
        Server enforces that VERIFIED, SUPERSEDED, or ARCHIVED records are not
        casually edited via this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePassportRecordRequest'
      responses:
        '200':
          description: Passport record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportRecord'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/InvalidStateTransitionError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /passport-records/{recordId}/submit:
    parameters:
      - $ref: '#/components/parameters/RecordId'
    post:
      tags: [Verification]
      summary: Submit passport record for verification
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitRecordRequest'
      responses:
        '200':
          description: Record submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportRecord'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/InvalidStateTransitionError'
  /passport-records/{recordId}/verify:
    parameters:
      - $ref: '#/components/parameters/RecordId'
    post:
      tags: [Verification]
      summary: Verify submitted passport record
      description: |
        Server-enforced checks include role, scope, lodge policy, and valid prior state
        (SUBMITTED or NEEDS_CLARIFICATION).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRecordRequest'
      responses:
        '200':
          description: Record verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportRecord'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/InvalidStateTransitionError'
  /passport-records/{recordId}/reject:
    parameters:
      - $ref: '#/components/parameters/RecordId'
    post:
      tags: [Verification]
      summary: Reject submitted passport record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRecordRequest'
      responses:
        '200':
          description: Record rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportRecord'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/InvalidStateTransitionError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /passport-records/{recordId}/clarification:
    parameters:
      - $ref: '#/components/parameters/RecordId'
    post:
      tags: [Verification]
      summary: Request clarification for a submitted record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClarificationRequest'
      responses:
        '200':
          description: Clarification requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportRecord'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/InvalidStateTransitionError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /passport-records/{recordId}/override:
    parameters:
      - $ref: '#/components/parameters/RecordId'
    post:
      tags: [Verification]
      summary: Override prior decision (audited)
      description: Lodge Mentor or exceptional admin path only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OverrideDecisionRequest'
      responses:
        '200':
          description: Decision overridden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportRecord'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/InvalidStateTransitionError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /verification-queue:
    get:
      tags: [Verification]
      summary: List pending verification queue
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [ASSIGNED, LODGE, DISTRICT]
            default: ASSIGNED
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PassportRecordStatus'
        - name: staleOnly
          in: query
          schema: { type: boolean, default: false }
        - name: lodgeId
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Verification queue items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVerificationQueueList'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MemberId:
      name: memberId
      in: path
      required: true
      schema: { type: string }
...
    RecordId:
      name: recordId
      in: path
      required: true
      schema: { type: string }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
  responses:
    UnauthenticatedError:
      description: Authentication is required or token is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication required.
              details: null
    ForbiddenError:
      description: Authenticated but not authorized for requested scope/action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action.
              details: null
    NotFoundError:
      description: Requested resource was not found in caller scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found.
              details: null
    ValidationError:
      description: Request body or parameters failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: VALIDATION_ERROR
              message: One or more validation errors occurred.
              details:
                fields:
                  eventDate: Must be a valid date.
    InvalidStateTransitionError:
      description: Requested workflow transition is not valid from current state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: INVALID_STATE_TRANSITION
              message: Record cannot be verified from current status.
              details:
                status: DRAFT
  schemas:
    RoleCode:
      type: string
      enum:
        - BROTHER
        - PERSONAL_MENTOR
        - LODGE_MENTOR
        - LODGE_REVIEWER
        - LODGE_ADMIN
        - DISTRICT_MENTOR
        - DISTRICT_ADMIN
        - SYSTEM_ADMIN
...
    PassportRecordStatus:
      type: string
      enum:
        - DRAFT
        - SUBMITTED
        - NEEDS_CLARIFICATION
        - VERIFIED
        - REJECTED
        - SUPERSEDED
        - ARCHIVED
...
    ItemType:
      type: string
      enum:
        - LEARNING_OUTCOME
        - MENTOR_SESSION
        - VISITATION
        - RITUAL_PARTICIPATION
        - MILESTONE
        - ANNUAL_COMMUNICATION
        - PREPARING_FOR_OFFICE
        - CUSTOM_SUPPLEMENT
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_ERROR
                - INVALID_STATE_TRANSITION
                - CONFLICT
                - FEATURE_DISABLED
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
            details:
              nullable: true
...
    UserProfile:
      type: object
      required: [id, email, displayName, status]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        mobileNumber: { type: string, nullable: true }
        givenName: { type: string, nullable: true }
        familyName: { type: string, nullable: true }
        displayName: { type: string }
        status: { type: string }
        lastLoginAt: { type: string, format: date-time, nullable: true }
...
    AuthTokenResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
    AuthSessionResponse:
      allOf:
        - $ref: '#/components/schemas/AuthTokenResponse'
        - type: object
          required: [user, roles, scopes]
          properties:
            user:
              $ref: '#/components/schemas/UserProfile'
            roles:
              type: array
              items:
                $ref: '#/components/schemas/RoleCode'
            scopes:
              type: object
              properties:
                lodges:
                  type: array
                  items: { type: string }
                districts:
                  type: array
                  items: { type: string }
                assignedMemberIds:
                  type: array
                  items: { type: string }
            featureFlags:
              type: array
              items: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
...
    PassportRecord:
      type: object
      required:
        [id, memberProfileId, districtId, lodgeId, sectionTemplateId, templateItemId, status, currentVersion, isOfficialProgress, createdAt, updatedAt]
      properties:
        id: { type: string }
        memberProfileId: { type: string }
        districtId: { type: string }
        lodgeId: { type: string }
        sectionTemplateId: { type: string }
        templateItemId: { type: string }
        templateIsDistrictCore: { type: boolean }
        templateLodgeId: { type: string, nullable: true }
        note: { type: string, nullable: true }
        eventDate: { type: string, format: date, nullable: true }
        reviewReason: { type: string, nullable: true }
        supersedesRecordId: { type: string, nullable: true }
        status:
          $ref: '#/components/schemas/PassportRecordStatus'
        currentVersion: { type: integer }
        isOfficialProgress: { type: boolean }
        submittedAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...
    VerificationQueueItem:
      type: object
      required: [id, passportRecordId, memberProfileId, lodgeId, currentStatus, submittedAt, isStale]
      properties:
        id: { type: string }
        passportRecordId: { type: string }
        memberProfileId: { type: string }
        assignedVerifierUserId: { type: string, nullable: true }
        lodgeId: { type: string }
        currentStatus:
          $ref: '#/components/schemas/PassportRecordStatus'
        submittedAt: { type: string, format: date-time }
        staleAfterAt: { type: string, format: date-time, nullable: true }
        isStale: { type: boolean }
...
    BrotherPassportSummary:
      type: object
      required: [memberProfileId, sections]
      properties:
        memberProfileId: { type: string }
        sections:
          type: array
          items:
            $ref: '#/components/schemas/BrotherPassportSectionSummary'
    BrotherPassportSectionSummary:
      type: object
      required: [sectionCode, sectionName, progressState]
      properties:
        sectionCode: { type: string }
        sectionName: { type: string }
        progressState:
          type: string
          enum: [NOT_STARTED, IN_PROGRESS, VERIFIED]
        latestStatus:
          $ref: '#/components/schemas/PassportRecordStatus'
        lastActivityAt: { type: string, format: date-time, nullable: true }
        pendingAction:
          type: string
          nullable: true
          enum: [SUBMIT_DRAFT, AWAITING_REVIEW, RESPOND_TO_CLARIFICATION]
...
    CreatePassportRecordRequest:
      type: object
      required: [districtId, lodgeId, sectionTemplateId, templateItemId]
      properties:
        districtId: { type: string }
        lodgeId: { type: string }
        sectionTemplateId: { type: string }
        templateItemId: { type: string }
        eventDate: { type: string, format: date, nullable: true }
        note: { type: string, nullable: true }
    UpdatePassportRecordRequest:
      type: object
      properties:
        eventDate: { type: string, format: date, nullable: true }
        note: { type: string, nullable: true }
    SubmitRecordRequest:
      type: object
      properties:
        note: { type: string, nullable: true }
    VerifyRecordRequest:
      type: object
      properties:
        note: { type: string, nullable: true }
    RejectRecordRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    ClarificationRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    OverrideDecisionRequest:
      type: object
      required: [targetStatus, reason]
      properties:
        targetStatus:
          type: string
          enum: [VERIFIED, REJECTED, SUPERSEDED]
        reason: { type: string }
...
    PaginatedPassportRecordList:
      type: object
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PassportRecord'
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
    PaginatedVerificationQueueList:
      type: object
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VerificationQueueItem'
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
...